fix(ai/collab): address Devin Review round 2 findings on PR #16
BUG-0001: ESRGAN tiling no longer leaves a black border on output.
The previous code unconditionally cropped 32 upscaled pixels from
every side of every tile, which made image-boundary tiles drop
pixels that no neighbouring tile could fill. The crop is now
position-aware: each side is only cropped when there is a real
neighbouring tile to contribute that overlap region. Per-tile crop
math is extracted into the testable `esrgan_tile_crop` helper with
new unit tests covering first-tile, last-tile, middle-tile, and
single-tile-fits-in-step cases.

BUG-0002 + ANALYSIS-0005: CRDT classifier no longer
misclassifies spot_color_remove (and any other future _remove /
_delete suffix) as a node Delete. The old suffix wildcards in
`is_delete_command` shadowed `is_document_setting_command` and
caused concurrent spot-color edits to be silently dropped under the
delete-wins rule. The classifier is now an explicit allow-list;
`spot_color_remove` is routed to DocumentSetting alongside the
upsert / catalog-load variants. A regression test in
crdt_merge.rs locks in the correct classification.

ANALYSIS-0001: SAM scale factor now uses (NET as f32) directly
rather than the `f32::from(NET as u16)` chain; lossless for
NET=1024 and robust to future increases up to 2^24.

Also: wire kcreate_core marketplace module into the bridge as
`DocumentBridgeError::Marketplace` (#[from]) so the WIP
template_install_local / template_remove entry points propagate
structured errors instead of stringifying them.

+++ crates/kcreate_ai/src/segment.rs
@@ -455,8 +455,10 @@ fn run_onnx_sam(
 
     // SAM keeps the aspect ratio by resizing the longer side to 1024
     // and zero-padding the shorter side. We compute the same here.
+    // `NET as f32` is lossless because NET (1024) is far below the
+    // f32 integer-precision limit of 2^24.
     let longer = width.max(height) as f32;
-    let scale = f32::from(NET as u16) / longer;
+    let scale = (NET as f32) / longer;
     let resized_w = ((width as f32) * scale).round() as u32;
     let resized_h = ((height as f32) * scale).round() as u32;
